Ship in both esm and cjs with this strat: https://github.com/garronej/ts-ci?tab=readme-ov-file#esm-for-bundlers-browser--cjs-for-node
