+++ examples/tanstack-router-file-based/package.json
@@ -29,7 +29,6 @@
         "eslint-plugin-react-hooks": "5.2.0",
         "eslint-plugin-react-refresh": "0.4.19",
         "typescript": "5.7.2",
-        "vite": "7.1.3",
-        "vite-plugin-commonjs": "0.10.4"
+        "vite": "7.1.3"
     }
 }
