+++ src/core/createOidc.ts
@@ -12,12 +12,7 @@ import { createEvtIsUserActive } from "./evtIsUserActive";
 import { createStartCountdown } from "../tools/startCountdown";
 import { toHumanReadableDuration } from "../tools/toHumanReadableDuration";
 import { toFullyQualifiedUrl } from "../tools/toFullyQualifiedUrl";
-import {
-    OidcInitializationError,
-    createFailedToFetchTokenEndpointInitializationError,
-    createIframeTimeoutInitializationError,
-    createWellKnownOidcConfigurationEndpointUnreachableInitializationError
-} from "./OidcInitializationError";
+import { OidcInitializationError } from "./OidcInitializationError";
 import { type StateData, generateStateUrlParamValue, STATE_STORE_KEY_PREFIX } from "./StateData";
 import { notifyOtherTabsOfLogout, getPrOtherTabLogout } from "./logoutPropagationToOtherTabs";
 import { notifyOtherTabsOfLogin, getPrOtherTabLogin } from "./loginPropagationToOtherTabs";
@@ -512,7 +507,9 @@ export async function createOidc_nonMemoized<
                             assert(error instanceof Error, "741947");
 
                             if (error.message === "Failed to fetch") {
-                                return createFailedToFetchTokenEndpointInitializationError({
+                                return (
+                                    await import("./diagnostic")
+                                ).createFailedToFetchTokenEndpointInitializationError({
                                     clientId,
                                     issuerUri
                                 });
@@ -682,18 +679,20 @@ export async function createOidc_nonMemoized<
                 if (result_loginSilent.outcome === "failure") {
                     switch (result_loginSilent.cause) {
                         case "can't reach well-known oidc endpoint":
-                            return createWellKnownOidcConfigurationEndpointUnreachableInitializationError(
+                            return (
+                                await import("./diagnostic")
+                            ).createWellKnownOidcConfigurationEndpointUnreachableInitializationError({
+                                issuerUri
+                            });
+                        case "timeout":
+                            return (await import("./diagnostic")).createIframeTimeoutInitializationError(
                                 {
-                                    issuerUri
+                                    redirectUri: homeUrlAndRedirectUri,
+                                    clientId,
+                                    issuerUri,
+                                    noIframe
                                 }
                             );
-                        case "timeout":
-                            return createIframeTimeoutInitializationError({
-                                redirectUri: homeUrlAndRedirectUri,
-                                clientId,
-                                issuerUri,
-                                noIframe
-                            });
                     }
 
                     assert<Equals<typeof result_loginSilent.cause, never>>(false);
@@ -715,7 +714,9 @@ export async function createOidc_nonMemoized<
                     assert(error instanceof Error, "433344");
 
                     if (error.message === "Failed to fetch") {
-                        return createFailedToFetchTokenEndpointInitializationError({
+                        return (
+                            await import("./diagnostic")
+                        ).createFailedToFetchTokenEndpointInitializationError({
                             clientId,
                             issuerUri
                         });
