+++ tsconfig.json
@@ -13,7 +13,9 @@
         "strict": true,
         "jsx": "react-jsx",
         "noFallthroughCasesInSwitch": true,
-        "skipLibCheck": true
+        "skipLibCheck": true,
+        // NOTE: Only for not getting error while developing, will be overwritten
+        "module": "es2020"
     },
     "include": ["src"]
 }
